fix: 추천목록레이어 이벤트 생성, 삭제 방식 변경
