+++ .eslintrc.json
@@ -7,7 +7,8 @@
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaVersion": 2020,
-    "sourceType": "module"
+    "sourceType": "module",
+    "@typescript-eslint/no-unused-vars": "off"
   },
   "plugins": ["@typescript-eslint"],
   "rules": {
