+++ client/scss/pages/home.scss
@@ -54,12 +54,11 @@ header {
 
       li {
         width: 80%;
+        height: 5%;
         padding: 0.5em;
         border-radius: 6px;
-        box-shadow: rgba(0, 0, 0, 0.07) 0px 1px 2px,
-          rgba(0, 0, 0, 0.07) 0px 2px 4px, rgba(0, 0, 0, 0.07) 0px 4px 8px,
-          rgba(0, 0, 0, 0.07) 0px 8px 16px, rgba(0, 0, 0, 0.07) 0px 16px 32px,
-          rgba(0, 0, 0, 0.07) 0px 32px 64px;
+        box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 12px;
+        background-color: rgb(247, 247, 247);
         display: flex;
         justify-content: space-around;
         align-items: center;
